arm: Added IR_IICAST

+++ test2/test.c
@@ -1,5 +1,5 @@
 
-int gl = 4;
+short gl = 4;
 
 int main(int argc) {
    return gl;
